Rename module to pauli.
