+++ pennylane/pauli/graph_colouring.py
