+++ tests/pauli/test_pauli_group.py
@@ -17,7 +17,7 @@ Unit tests for the :mod:`pauli_group`  functions in ``groups/pauli_group.py``.
 import pytest
 from pennylane import Identity, PauliX, PauliY, PauliZ
 
-from pennylane.groups.pauli import pauli_group, pauli_mult, pauli_mult_with_phase
+from pennylane.pauli.pauli import pauli_group, pauli_mult, pauli_mult_with_phase
 
 
 class TestPauliGroup:
